Organize services in designated namespace
