+++ site/src/Librecores/ProjectRepoBundle/Resources/config/legacy_aliases.yml
@@ -2,7 +2,7 @@ services:
         update_project_info_service: '@Librecores\ProjectRepoBundle\Consumer\UpdateProjectInformation'
         librecores.util.markuptohtmlconverter:  '@Librecores\ProjectRepoBundle\Util\MarkupToHtmlConverter'
         librecores.util.GitHubApiService: '@Librecores\ProjectRepoBundle\Service\GitHub\GitHubApiService'
-        librecores.util.queuedispatcherservice: '@Librecores\ProjectRepoBundle\Util\QueueDispatcherService'
+        librecores.util.queuedispatcherservice: '@Librecores\ProjectRepoBundle\Service\QueueDispatcherService'
         app.form.type.project: '@Librecores\ProjectRepoBundle\Form\Type\ProjectType'
         app.form.type.organization: '@Librecores\ProjectRepoBundle\Form\Type\OrganizationType'
         librecores.form_extention.icon: '@Librecores\ProjectRepoBundle\Form\Extension\ButtonTypeIconExtension'
